Add normalize support to NPEC (enabled by default)
